passing data bottom-up to parent component
To send data up to the parent component, roughly the opposite of props,
you should declare a function in the parent component.
Then, pass that function to the child component as a prop.
In the child component, you have to call the function inside a listener
and pass any data, if necessary, as argument. the function will be triggered in the
parent component and use the data the child component send.