add conditional rendering
There are couple different ways to create a conditional rendering. You can store the content in a variable
and assign the variable accordingly to the content. Or use an if statemente to return different
content for each situation. It is important to keep the component lean and the code easy to understand.